% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolgenisSession.R
\docType{class}
\name{MolgenisSession-class}
\alias{MolgenisSession-class}
\title{Class MolgenisSession.}
\arguments{
\item{host}{Hostname of the MOLGENIS server}

\item{username}{username used to authenticate}

\item{password}{password used to authenticate}
}
\description{
A Session on a MOLGENIS server.
}
